refactor(contract): move shared request and auth plumbing into api/common.rs
